fix: changed array to long type, correct solution but inefficient

+++ hackerrank/ArrayManipulation.java
@@ -26,7 +26,7 @@ public class ArrayManipulation {
 
 	static long arrayManipulation(int n, int[][] queries) {
         long max = 0;
-        int[] arr = new int[n]; 
+        long[] arr = new long[n]; 
         for (int i = 0; i < queries.length; i++){
             for (int j = queries[i][0]-1; j <= queries[i][1]-1; j++){
                 arr[j] += queries[i][2];
